Refactoring amplifyUtil.ts into three files amplifyUtils.ts, fileUtils.ts, and transformConfig.ts. for better ease of reuse and logical consistency.
