+++ packages/graphql-transformer-core/src/index.ts
@@ -8,11 +8,13 @@ import { stripDirectives } from './stripDirectives'
 import {
     buildProject as buildAPIProject,
     uploadDeployment as uploadAPIProject,
-    readSchema as readProjectSchema,
     migrateAPIProject,
     revertAPIMigration,
-    readProjectConfiguration
 } from './util/amplifyUtils'
+import {
+    readSchema as readProjectSchema,
+    loadProject as readProjectConfiguration
+} from './util/transformConfig'
 
 export * from './errors'
 export * from './util'
